ReciPie/pyproject.toml
[project]
name = "recipie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "jinja2>=3.1.6",
    "langchain>=1.2.7",
    "langchain-groq>=1.1.1",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "uvicorn[standard]>=0.40.0",
]
